// src/enums/offline.ts
export enum OFFLINE_METHOD {
  POST = 'POST',
  GET = 'GET',
  PUT = 'PUT',
  PATCH = 'PATCH',
  DELETE = 'DELETE',
}

export enum OFFLINE_ACTIONS {
  CREATE_ASSET = 'CREATE_ASSET',
  EDIT_ASSET = 'EDIT_ASSET',
  CREATE_POINT = 'CREATE_POINT',
  EDIT_POINT = 'EDIT_POINT',
  DELETE_POINT = 'DELETE_POINT',
  CREATE_NOTE = 'CREATE_NOTE',
  ADD_ASSET_FILES = 'ADD_ASSET_FILES',
  CREATE_ASSIGNMENT_FOOLDER = 'CREATE_ASSIGNMENT_FOOLDER',
  UPDATE_ASSIGNMENT_FOOLDER = 'UPDATE_ASSIGNMENT_FOOLDER',
  DELETE_ASSIGNMENT_FOOLDER = 'DELETE_ASSIGNMENT_FOOLDER',
  CREATE_ASSIGNMENT_PANEL = 'CREATE_ASSIGNMENT_PANEL',
  EDIT_WO = 'EDIT_WO',
  EDIT_WO_TECH = 'EDIT_WO_TECH',
}